OTAT-468: Rename package
