+++ src/main/java/org/flickit/flickitassessmentcore/adapter/in/rest/exception/common/ValidationError.java
@@ -1,4 +1,4 @@
-package org.flickit.flickitassessmentcore.adapter.in.rest.advice.common;
+package org.flickit.flickitassessmentcore.adapter.in.rest.exception.common;
 
 public record ValidationError(String propertyPath,
                               String message) {
